Add aria-label to social media links
Fixes https://www.pivotaltracker.com/story/show/169552905
and https://www.pivotaltracker.com/story/show/169552982

This adds an an aria-label to social media links as well as aria-hidden to the svgs themselves. This does not address Pinterest accessibility because there seems to be an issue with how the Pinterest library dynamically removes the link's href, rendering it inaccessible. There is a separate story to address that here: https://www.pivotaltracker.com/story/show/169690936