Introduce All products button text setting
As long as this button text is set, whenever the total number of products exceeds the number of products featured on the Home page, the button will appear. This also fixes the bug where pagination was inadvertently displaying on the Home page, and cleans up some button styles as well.